backend/requirements.txt
flask
flask-cors
openai-whisper
gtts